requirements.txt
Django>=4.2.1,<4.3.0
psycopg2-binary>=2.9.1,<2.10.0
djangorestframework>=3.14.0,<3.15.0
